build: move string_view to its own library
