+++ vowpalwabbit/ccb_label.cc
@@ -18,7 +18,7 @@
 #include "reductions/interactions.h"
 #include "vw.h"
 #include "vw_math.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <algorithm>
 #include <cmath>
