+++ vowpalwabbit/config/options_cli.cc
@@ -7,7 +7,7 @@
 #include "config/option.h"
 #include "text_utils.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <fmt/format.h>
