+++ vowpalwabbit/cost_sensitive.cc
@@ -11,7 +11,7 @@
 #include "shared_data.h"
 #include "vw.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <cfloat>
 #include <cmath>
