+++ vowpalwabbit/global_data.cc
@@ -21,7 +21,7 @@
 #include "shared_data.h"
 #include "vw_allreduce.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <rapidjson/stringbuffer.h>
 #include <rapidjson/writer.h>
