+++ vowpalwabbit/label_type.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "future_compat.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <cstdint>
