+++ vowpalwabbit/multiclass.cc
@@ -10,7 +10,7 @@
 #include "shared_data.h"
 #include "vw.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <climits>
 #include <cstring>
