+++ vowpalwabbit/named_labels.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "vw_fwd.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <string>
 #include <unordered_map>
