+++ vowpalwabbit/parse_args.h
@@ -6,7 +6,7 @@
 #include "config/options.h"
 #include "text_utils.h"
 #include "vw_fwd.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 // Used in parse_source
 struct input_options
