+++ vowpalwabbit/parse_primitives.cc
@@ -6,7 +6,7 @@
 
 #include "hash.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <algorithm>
 #include <cctype>
