+++ vowpalwabbit/parser.h
@@ -24,7 +24,7 @@
 #include "object_pool.h"
 #include "queue.h"
 #include "vw_fwd.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <atomic>
 #include <memory>
