+++ vowpalwabbit/reductions/automl.h
@@ -11,7 +11,7 @@
 #include "rand_state.h"
 #include "scored_config.h"
 #include "vw_fwd.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 #include <fmt/format.h>
