+++ vowpalwabbit/reductions/cb/cb_adf.cc
@@ -14,7 +14,7 @@
 #include "shared_data.h"
 #include "vw.h"
 #include "vw_exception.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 #include "vw_versions.h"
 
 #undef VW_DEBUG_LOG
