+++ vowpalwabbit/simple_label.cc
@@ -9,7 +9,7 @@
 #include "io/logger.h"
 #include "parse_primitives.h"
 #include "vw.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 #include "vw_string_view_fmt.h"
 
 #include <cfloat>
