+++ vowpalwabbit/tag_utils.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "vw_fwd.h"
-#include "vw_string_view.h"
+#include "vowpalwabbit/vw_string_view.h"
 
 namespace VW
 {
